Fix layer init functions kwargs getting overwritten
There were a few functions of the form:
foo(rng::AbstractRNG; kwargs...) = (dims...; kwargs...) -> foo(rng, dims...; kwargs...)

The intention was that you could do `foo(my_kw=42)` to get a callable, and then later
call that callable with the kwarg already set. However, the `kwargs` variable of the
lambda shadowed the original `kwargs`. This is a fix.